% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{compute_diff}
\alias{compute_diff}
\title{The difference between the estimated and real dimensionality reduction space.}
\usage{
compute_diff(
  W_esti,
  d,
  true_B = B,
  type = "matrix-dot",
  D.training = D.training
)
}
\arguments{
\item{W_esti, }{estimated reduced dimension space.}

\item{d, }{true order of the reduced dimension space.}

\item{true_B, }{true reduced dimension space.}

\item{type, }{method to calculate the difference, contain 'matrix-dot', 'e-norm', 'R-sqare'.}

\item{D.training, }{datasets consists source and target data, refer <glmtrans> package.}
}
\description{
The difference between the calculated and real dimensionality reduction space.
}
