% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{lasso_w}
\alias{lasso_w}
\title{sufficient dimension reduction with original lasso method.}
\usage{
lasso_w(D.training.origin, target_Y)
}
\arguments{
\item{D.training.origin, }{datasets consists source and target data, refer <glmtrans> package.}

\item{target_Y, }{transformed target response, see <transform_y> function.}
}
\description{
traditional method to get dimension reduct space.
}
