% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{transfer_sdr}
\alias{transfer_sdr}
\title{sufficient dimension reduction with transfer learning.}
\usage{
transfer_sdr(D.training.origin, source_Y, target_Y)
}
\arguments{
\item{D.training.origin, }{datasets consists source and target data, refer <glmtrans> package.}

\item{source_Y, }{transformed source response, see <transform_y> function.}

\item{target_Y, }{transformed target response, see <transform_y> function.}
}
\description{
using transfer learning to get dimension reduct space.
}
