% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{transform_y}
\alias{transform_y}
\title{response's transformation}
\usage{
transform_y(D.training, K = 1, type = "bs", df = 5, degree = 1)
}
\arguments{
\item{D.training, }{datasets consists source and target data, refer <glmtrans> package.}

\item{K, }{the number of source datasets.}

\item{type, }{response types, 'bs' means B-splines, 'slicing' means slicing.}

\item{df, }{number of transform directions.}

\item{degree, }{used in bs method.}
}
\description{
response's transformation
}
